Code clean up
Removed the println's that I used to fix.  Added a todo at the error
site.  Found out that someone else had already found out the
JSON.getLong call was causing problems and fixed it in the
resolvVanityUrl method but DIDN'T BOTHER TO FIX IT OTHER PLACES *RAGE*.
Added a todo at that location as well.