Worked on making recursive
Started working on making the player fetching recursive.  It is now
recursive, but I've seemed to have broken it.  Commit now just to get a
'safe' spot.  I should have committed early when I was something that
was a little better (but still not great).  As an aside, our credentials
have change, due to my blunder.