Added more PSQL
***I added more PSQL code.  I did very fast and dirty testing of the
retrieveOwnedApplication method, consisting of checking the first five
returned games that it said I owned (which I did) and then checked the
bounds (I own 82 games and could access 0-81 in the returned list; the
83th game threw an out of bound exception.  So it probabilistically
returned my data).
*** One issue I encountered was with UpdateEntity.  Given the data I
have, I cannot know where to split the options between the Set [...] and
the Where [...].  This issue can be avoided by simply including two
large strings, one corresponding to the SET and the second to the WHERE
clauses, but this is a work around.
***  I still have some concern about attempting to make these SQL in
this very modular style, as it pushes a lot of knowledge about the
database's internal workings on the Controller (to be able to form some
of the strings).  I still wonder if making more specified methods which
all get resolved to the one call in the view may be more appropriate.
That way the Controller can just feed in whatever variables it wants and
the Model has to resolve it correctly.  This would also allow for more
complex SQL.
***Another topic is a possible solution to the issue of table column
headers.  If we do go the route of more specialized methods and complex
SQL, we can fetch all the columns as {appropriate column name} and then
include that as the first row in the resulting Object[][].  This would
mean that whenever a fetch was returned, the corresponding column
headers are wrapped with it.
***Also, is this class going to stay static, as it probably needs to be
instantiated so that it can properly set up the connection, otherwise it
needs to have a static Setup and Teardown functions.