Finished Populate
Finally got around to added support for achievements.  Adds a set number
of games to check for achievements and then for each player checks their
achievements and adds ones they have gotten.  Also set up code to allow
for a total data mine, where all games will be checked for achievements
and all achievements logged.  It is highly advised that that is only
used in dire situations.