Catching Exceptions!
Insert and update can now intelligently tell you if you've done one of
two things wrong - 1, tried to insert a different type into a column
(i.e. a string into an int attribute), and if the foreign key does not
exist. Might not handle putting an int into a string attribute. Might be
cool with that?