Fixed issue with Inserting on filtered data
There was an ArrayOutOfBounds exception thrown when inserting on data
that was filtered.  Turns out the rowsChanged was never being emptied,
so previously changed data was still trying to be changed, and when the
data set was filtered it lead to issues.  Atleast, that is what it seems
like; as emptying the rowsChanged after changing those rows fixed the
issue.